extracted CLI module
