+++ package.json
@@ -30,6 +30,6 @@
     "webpack-sources": "1.3.0"
   },
   "devDependencies": {
-    "mocha": "^5.2.0"
+    "mocha": "5.2.0"
   }
 }
