+++ src/cli.js
@@ -1,0 +1,126 @@
+
+var Msg = {
+    PrintHelp: 0,
+    PrintVersion: 1,
+    MinifyFile: 2
+}
+
+var defaultConfig = {
+    msg: Msg.MinifyFile,
+    inputFilePath: "dist/index.js",
+    extraRound: false,
+    replace: false
+}
+
+var fileReadWriteConfig = { encoding: "utf8" }
+
+var toKb = function (byteLength) {
+
+    return byteLength / 1000
+}
+
+var toString = function (a) {
+    return a + ""
+}
+
+var argvToConfig = function (argv) {
+
+    var cmdOrFilePath = argv[2]
+
+    switch (cmdOrFilePath) {
+
+        case "--help":
+            return { msg: Msg.PrintHelp }
+
+        case "--version":
+            return { msg: Msg.PrintVersion }
+
+        default:
+
+            var config = defaultConfig
+
+            config.inputFilePath = cmdOrFilePath
+            config.extraRound = argv.indexOf("--extraRound") !== -1
+            config.replace = argv.indexOf("--replace") !== -1
+
+            return config
+    }
+
+}
+
+var padString = function (text, length, toRight) {
+
+    var lengthDiff = length - text.length
+
+    if (lengthDiff > 0) {
+
+        return toRight
+            ? " ".repeat(lengthDiff).concat(text)
+            : text.concat(" ".repeat(lengthDiff))
+    }
+    else if (lengthDiff < 0) {
+
+        return text.slice(0, length)
+    }
+    else {
+
+        return text
+    }
+}
+
+var toResultString = function (srcPathSizeArray) {
+
+    var toResultEntry = function (srcPathSize) {
+
+        return "│ "
+            + padString(srcPathSize[0], 6, true) + " │ "
+            + padString(srcPathSize[1], 18, false) + " │ "
+            + padString(srcPathSize[2], 10, true) + " │"
+    }
+
+    return [
+        "┌────────┬────────────────────┬────────────┐",
+        "│ source │ rel path           │    kb size │",
+        "├────────┼────────────────────┼────────────┤",
+    ]
+        .concat(srcPathSizeArray.map(toResultEntry))
+        .concat([
+            "└────────┴────────────────────┴────────────┘"
+        ])
+        .join("\n")
+}
+
+var toHelpString = function (version) {
+
+    return [
+        "",
+        "   elm-minify " + version,
+        "",
+        "Usage:",
+        "",
+        "   elm-minify <input>",
+        "",
+        "<input>:                   Defaults to 'dist/index.js'",
+        "",
+        "   <filepath>.js           Minify to <filepath>.min.js",
+        "",
+        "       [--extra-round]     Run an additional round of compression",
+        "       [--replace]         Overwrite the input file with minified output",
+        "",
+        "   --version                       Show package version",
+        "   --help                          Show this help message",
+        ""
+    ].join("\n")
+}
+
+module.exports = {
+    Msg: Msg,
+    defaultConfig: defaultConfig,
+    toKb: toKb,
+    toString: toString,
+    argvToConfig: argvToConfig,
+    toHelpString: toHelpString,
+    toResultString: toResultString,
+    padString: padString,
+    fileReadWriteConfig: fileReadWriteConfig,
+}
